requirements.txt
anyio==4.1.0
certifi==2023.11.17
exceptiongroup==1.2.0
h11==0.14.0
httpcore==1.0.2
httpx==0.25.1
idna==3.4
psutil==5.9.6
python-dotenv==1.0.0
python-telegram-bot==20.6
schedule==1.2.1
sniffio==1.3.0
